+++ packages/moengage_flutter/moengage_flutter_ios/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Release Version
 - [minor] Added SPM integration support
+- [minor] Added support for initialization from `Info.plist` data
 
 # 14-10-2025